+++ src/r_draw.h
@@ -146,6 +146,7 @@ extern byte         *dc_translation;
 // No Spectre effect needed.
 void R_DrawSpan(void);
 void R_DrawColorSpan(void);
+void R_DrawNoTextureSpan(void);
 
 void R_InitBuffer(int width, int height);